--- a/src/rootUtils.h
+++ b/src/rootUtils.h
@@ -2,6 +2,7 @@
 #define PODIO_ROOT_UTILS_H // NOLINT(llvm-header-guard): internal headers confuse clang-tidy
 
 #include "podio/CollectionIDTable.h"
+#include "podio/utilities/MiscHelpers.h"
 #include "podio/utilities/RootHelpers.h"
 #include "podio/utilities/TypeHelpers.h"
 
@@ -278,25 +279,6 @@ inline auto reconstructCollectionInfo(TTree* eventTree, podio::CollectionIDTable
   }
 
   return collInfo;
-}
-
-/**
- * Sort the input vector of strings alphabetically, case insensitive.
- */
-inline std::vector<std::string> sortAlphabeticaly(std::vector<std::string> strings) {
-  // Obviously there is no tolower(std::string) in c++, so this is slightly more
-  // involved and we make use of the fact that lexicographical_compare works on
-  // ranges and the fact that we can feed it a dedicated comparison function,
-  // where we convert the strings to lower case char-by-char. The alternative is
-  // to make string copies inside the first lambda, transform them to lowercase
-  // and then use operator< of std::string, which would be effectively
-  // hand-writing what is happening below.
-  std::ranges::sort(strings, [](const auto& lhs, const auto& rhs) {
-    return std::lexicographical_compare(
-        lhs.begin(), lhs.end(), rhs.begin(), rhs.end(),
-        [](const auto& cl, const auto& cr) { return std::tolower(cl) < std::tolower(cr); });
-  });
-  return strings;
 }
 
 /**