--- a/tools/json-to-yaml
+++ b/tools/json-to-yaml
@@ -1,4 +1,5 @@
 #!/usr/bin/env python3
+"""Tiny script to ingest a json string and dump it as a yaml string"""
 
 import sys
 import json
@@ -6,6 +7,7 @@ import yaml
 
 
 def main():
+    """Main, read json from stdin and dump yaml to stdout"""
     input_data = sys.stdin.read()
     model_def = json.loads(input_data)
     print(yaml.dump(model_def, sort_keys=False, default_flow_style=False))