--- a/tools/json-to-yaml
+++ b/tools/json-to-yaml
@@ -1,0 +1,15 @@
+#!/usr/bin/env python3
+
+import sys
+import json
+import yaml
+
+
+def main():
+    input_data = sys.stdin.read()
+    model_def = json.loads(input_data)
+    print(yaml.dump(model_def, sort_keys=False, default_flow_style=False))
+
+
+if __name__ == "__main__":
+    main()