--- a/tools/src/argparseUtils.h
+++ b/tools/src/argparseUtils.h
@@ -2,9 +2,12 @@
 #define PODIO_TOOLS_ARGPARSEUTILS_H // NOLINT(llvm-header-guard): folder structure not suitable
 
 #include <algorithm>
+#include <charconv>
 #include <iostream>
+#include <ranges>
 #include <stdexcept>
 #include <string>
+#include <string_view>
 #include <vector>
 
 template <typename... Flags>
@@ -12,30 +15,21 @@ auto findFlags(const std::vector<std::string>& argv, Flags... flags) {
   return std::find_if(argv.begin(), argv.end(), [&](const auto& elem) { return ((elem == flags) || ...); });
 }
 
-inline std::vector<std::string> splitString(const std::string& str, const char delimiter) {
-  std::vector<std::string> tokens;
-  std::string token;
-  for (char ch : str) {
-    if (ch == delimiter) {
-      if (!token.empty()) {
-        tokens.push_back(token);
-        token.clear();
-      }
-    } else {
-      token += ch;
-    }
+inline std::vector<std::string_view> splitString(const std::string& str, const char delimiter) {
+  std::vector<std::string_view> tokens;
+
+  for (const auto& token : std::ranges::views::split(str, delimiter)) {
+    tokens.emplace_back(token.begin(), token.end());
   }
-  if (!token.empty()) {
-    tokens.push_back(token);
-  }
+
   return tokens;
 }
 
-inline size_t parseSizeOrExit(const std::string& str) {
+inline size_t parseSizeOrExit(std::string_view str) {
   try {
-    std::size_t pos{};
-    const auto number = std::stoull(str, &pos);
-    if (pos != str.size()) {
+    std::size_t number{};
+    const auto [ptr, err] = std::from_chars(str.data(), str.data() + str.size(), number);
+    if (err != std::errc{} || ptr != str.data() + str.size()) {
       throw std::invalid_argument("");
     }
     return number;