--- a/tools/src/argparseUtils.h
+++ b/tools/src/argparseUtils.h
@@ -4,7 +4,6 @@
 #include <algorithm>
 #include <charconv>
 #include <iostream>
-#include <stdexcept>
 #include <string>
 #include <string_view>
 #include <vector>
@@ -15,7 +14,6 @@ auto findFlags(const std::vector<std::string>& argv, Flags... flags) {
 }
 
 inline size_t parseSizeOrExit(std::string_view str) {
-  std::cout << "Parsing '" << str << "' into a size_t" << std::endl;
   std::size_t number{};
   const auto [ptr, err] = std::from_chars(str.data(), str.end(), number);
   if (err != std::errc{} || ptr != str.end()) {