BEGINRELEASENOTES
- Replace the `podio-dump` python implementation by a `podio-dump` wrapper script around the new `podio-dump-tool` c++ implementation. **This is an (almost) drop-in replacement**, only legacy files (i.e. files written before the introduction of the `Frame`) cannot be read by this.
  - The existing python implementation is still available as `podio-dump-legacy`. 
- **podio now depends on the [*fmtlib*](https://github.com/fmtlib/fmt) formatting library** (version 9 or greater) as the `podio-dump-tool` uses that for formatting.
  - This might change again in the future, but compilers don't have enough support for `std::format` yet to easily replace this.

ENDRELEASENOTES

This has taken a bit to develop (see below for the original description). The way it is now is a combination of two tools that are wrapped in a thin bash script that acts as the entry point and is called `podio-dump`. This is effectively only necessary because the datamodel definitions are stored as JSON inside the file and it's just way easier to transform that back to YAML in python than it is in c++. Hence, the bash script only checks for the corresponding flag and if that is present pipes the output of the c++ `podio-dump-tool` to a python script (`json-to-yaml`). Otherwise it simply calls the `podio-dump-tool` passing on all the arguments.

The dependency on *fmtlib* is necessary because it supports formatting of ranges and easy opt-in for implementations via `operator<<`. Additionally, `std::print` (and `println`) are only scheduled for c++23. Once there is more / all necessary support in `<format>`, adjusting to that and dropping the fmtlib dependency should be possible.

### Original description

This is an attempt at making `podio-dump` quicker after several complaints (e.g. https://github.com/key4hep/EDM4hep/issues/312). After some "profiling" it turns out that the slowest part in the python implementation is the loop over all the collections which can be significantly sped up by going to c++. In my local timings the current (python based) `podio-dump` is almost ten times slower than this (c++ based) `podio-dump-cpp`) for dumping the `example_frame.root` file from the tests (times via `time`)

| | `podio-dump` | `podio-dump-cpp` |
|--|--|--| 
| real | 12.393s |1.513s |
| user |8.522s | 1.251s|
| sys | 3.823s| 0.296s |

The main disadvantages of  the c++ implementation are that we need quite a bit of boilerplate for things that are trivial in python, e.g.:
- We have to manually implement argument parsing and (parts of) the `tabulate` functionality
  - Since formatting with `iostream` and `iomanip` is bordering on masochism, I have decided to pull in [`fmt`](https://fmt.dev/latest/index.html) for now. In principle c++20 has similar functionality in `std::format` (but no `fmt::print` that only comes with c++26). However, that requires gcc >=13 and clang >= 16.
- Dumping datamodel definitions in YAML is missing entirely at the moment, since that would require dumping the internal json format as YAML. In python this is literally these 2 lines:
 https://github.com/AIDASoft/podio/blob/d275460be9c098f2a1d83e67ec63398538673d05/tools/podio-dump#L99-L100
 in c++ this would require at least one other library to be pulled in
 
Since dumping the datamodel would require quite a bit of work in c++, I would be in favor of keeping that in python in a separate tool, while the other functionality could be covered by the c++ implementation.

TODO:
- [x] Implement detailed dumping mode
- [x] Builds on top of #618 
- [x] Includes #621