@Zehvogel 2024-06-21
I wonder how an RDataFrame-based python version (with pre-compiled functions) would fare on a performance vs. comfort scale