@tmadlener 2025-04-01
I have rebased this onto the latest version of podio. The main difference to before is that `podio-dump` is now a thin bash script that simply wraps the `podio-dump-tool` c++ executable. In case the user wants to dump an EDM via `--dump-edm` we invoke a thin python tool `json-to-yaml` that does the conversion on the fly. Now this is a drop-in replacement for the previous version of `podio-dump` implemented in python. The one thing the c++ implementation does not handle is reading pre-release legacy files. (I don't think that is an actual requirement any longer).

There are some formatting differences wrt to the python implmentation, which are pretty much all related to the alignment of numerical outputs in the tables.

Using ROOT 6.34.04 and python 3.11.11 I get the following times (simply using the test times here):

**Now**
```
1/7 Test #185: podio-dump-help ..................   Passed    0.03 sec
2/7 Test #186: podio-dump-root ..................   Passed    3.58 sec
3/7 Test #187: podio-dump-detailed-root .........   Passed    3.51 sec
4/7 Test #190: podio-dump-sio ...................   Passed    0.04 sec
5/7 Test #191: podio-dump-detailed-sio ..........   Passed    0.04 sec
6/7 Test #194: podio-dump-rntuple ...............   Passed    0.24 sec
7/7 Test #195: podio-dump-detailed-rntuple ......   Passed    0.23 sec

```
and
```
1/6 Test #160: datamodel_def_store_roundtrip_root ................   Passed    0.90 sec
2/6 Test #161: datamodel_def_store_roundtrip_root_extension ......   Passed    0.81 sec
3/6 Test #162: datamodel_def_store_roundtrip_sio .................   Passed    0.68 sec
4/6 Test #163: datamodel_def_store_roundtrip_sio_extension .......   Passed    0.60 sec
5/6 Test #164: datamodel_def_store_roundtrip_rntuple .............   Passed    0.81 sec
6/6 Test #165: datamodel_def_store_roundtrip_rntuple_extension ...   Passed    0.71 sec
```

**Before**
```
1/7 Test #185: podio-dump-help ..................   Passed    0.17 sec
2/7 Test #186: podio-dump-root ..................   Passed    8.50 sec
3/7 Test #187: podio-dump-detailed-root .........   Passed    9.35 sec
4/7 Test #190: podio-dump-sio ...................   Passed    8.08 sec
5/7 Test #191: podio-dump-detailed-sio ..........   Passed    7.71 sec
6/7 Test #194: podio-dump-rntuple ...............   Passed    8.36 sec
7/7 Test #195: podio-dump-detailed-rntuple ......   Passed    8.00 sec
```
and
```
1/6 Test #160: datamodel_def_store_roundtrip_root ................   Passed    2.62 sec
2/6 Test #161: datamodel_def_store_roundtrip_root_extension ......   Passed    2.52 sec
3/6 Test #162: datamodel_def_store_roundtrip_sio .................   Passed    2.32 sec
4/6 Test #163: datamodel_def_store_roundtrip_sio_extension .......   Passed    2.30 sec
5/6 Test #164: datamodel_def_store_roundtrip_rntuple .............   Passed    2.63 sec
6/6 Test #165: datamodel_def_store_roundtrip_rntuple_extension ...   Passed    2.53 sec
```

So, depending on which ROOT backend is used the speedup is somewhere between 2x and 30x. For SIO it is another factor 10 on top of that (for the test files). Even if the speedup is "only" a factor two for the most prevalent file format at the moment, I still think this would be worth it, also considering the increased implementation side, since many things that came packaged for python are now handrolled here.

----

I have checked `fmt`-lib down to version 8 which didn't compile, so minimal fmt version is 9. Things work unchanged up to version 11 (aka the latest at the moment).