@jmcarcell 2025-04-07
I don't have any more comments and probably won't have another look in detail unless I find something. I have been using this and it feels soooo much better to use, certainly way faster even in debug builds it is very fast. It also crashes faster when reading an incompatible file :smile: 
I wonder if having a better python implementation of the readers (I'm quite sure that's where most of the time was spent) would have helped. In any case this is good to have for a tool that is (should be) used a lot.