@jmcarcell 2025-04-10
This is now in the Key4hep nightlies on Alma 9 and Ubuntu 24. On Ubuntu 22, it doesn't build with GCC 11 anymore with multiple failures:
```
/podio/tools/src/podio-dump-tool.cpp:85:35: error: could not convert '{<expression error>}' from '<brace-enclosed initializer list>' to 'std::vector<long unsigned int>'
   85 |       return {parseSizeOrExit(*it)};
```

```
/podio/tools/src/podio-dump-tool.cpp:78:54: error: no match for 'operator|' (operand types are 'std::ranges::split_view<std::basic_string_view<char>, std::ranges::single_view<char> >' and 'std::ranges::views::__adaptor::_Partial<std::ranges::views::_Transform, parseEventRange(std::string_view)::<lambda(auto:32&&)> >')
   78 |     auto colonSplitRange = evtRange | rv::split(':') |
      |                            ~~~~~~~~~~~~~~~~~~~~~~~~~ ^
      |                                     |
      |                                     std::ranges::split_view<std::basic_string_view<char>, std::ranges::single_view<char> >
   79 |         rv::transform([](auto&& subrange) { return std::string_view(subrange.begin(), subrange.end()); });
      |         ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
      |                      |
      |                      std::ranges::views::__adaptor::_Partial<std::ranges::views::_Transform, parseEventRange(std::string_view)::<lambda(auto:32&&)> >
```

```
/podio/tools/src/podio-dump-tool.cpp:79:57: error: no matching function for call to 'std::basic_string_view<char>::basic_string_view(std::ranges::split_view<std::basic_string_view<char>, std::ranges::single_view<char> >::_InnerIter<false>, std::default_sentinel_t)'
   79 |         rv::transform([](auto&& subrange) { return std::string_view(subrange.begin(), subrange.end()); });
```

I wasn't aware of this, but actually building with GCC 11 (with tests) hasn't been possible since #626. I think probably the way to go forward with this is to request GCC > 11 rather than trying to make the errors go away.