@jmcarcell 2025-04-11
On the other side, GCC 15 is going to be released in ~ 1 month, so by now GCC 11 is already 4 years old. In addition no one seems to be building podio with tests since https://github.com/AIDASoft/podio/pull/626 is from last year's summer.